fix(project): create copryright component
